samples/agent/adk/adk_salesforce/pyproject.toml
[project]
name = "a2ui-adk-salesforce"
version = "0.1.0"
description = "Sample Google ADK-based Salesforce agent that uses a2ui extension."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
    "a2ui",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
google-adk = { git = "https://github.com/google/adk-python.git", rev = "143ad44" }
